# docs/openapi.yaml
openapi: 3.0.3
info:
  title: jjobb_v2 Graduate Network API
  version: 2.0.0-sprint4
  description: Sprint 4 recommendations, field trips, networking school scope (+ Sprint 2–3). v1 `/api` keys are preserved.
servers:
  - url: http://localhost:5000
    description: Local / Docker Compose backend
...
paths:
...
  /api/auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: current user }
        '401': { description: UNAUTHENTICATED }
...
  /api/schools:
    get:
      tags: [Schools]
      summary: Public list of active schools (signup)
      parameters:
        - in: query
          name: include_inactive
          schema: { type: boolean }
      responses:
        '200': { description: '{ schools: [] }' }
    post:
      tags: [Schools]
      security: [{ bearerAuth: [] }]
      summary: system_admin manage
      responses:
        '201': { description: created }
        '403': { description: FORBIDDEN }
...
  /api/me/permissions:
    get:
      tags: [IAM]
      security: [{ bearerAuth: [] }]
      summary: Role × menu × actions for current JWT
      responses:
        '200':
          description: '{ role, school_id, menus[] }'
  /api/audit-logs:
    get:
      tags: [IAM]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: '{ logs, pagination }' }
...
  /api/resumes:
    get:
      tags: [Resumes]
      security: [{ bearerAuth: [] }]
      summary: List own resumes (staff may pass user_id in same school)
      responses:
        '200': { description: '{ resumes: [] }' }
    post:
      tags: [Resumes]
      security: [{ bearerAuth: [] }]
      summary: Create resume with items (REQ-RSM-001/002)
      responses:
        '201': { description: created }
        '403': { description: FORBIDDEN }
...
  /api/notifications:
    get:
      tags: [Notify]
      security: [{ bearerAuth: [] }]
      summary: In-app notifications (REQ-MSG-010)
      responses:
        '200': { description: '{ notifications, unread_count }' }
  /api/counseling-journals/stats:
    get:
      tags: [Counseling]
      security: [{ bearerAuth: [] }]
      summary: Type/count/period stats (REQ-CNS-005)
      responses:
        '200': { description: '{ stats }' }
...
  /api/recommendations/me:
    get:
      tags: [Recommendations]
      security: [{ bearerAuth: [] }]
      summary: Ranked recommendations with reasons (REQ-REC-004)
      responses:
        '200': { description: '{ recommendations, scoring_factors }' }
  /api/recommendations/associated:
    get:
      tags: [Recommendations]
      security: [{ bearerAuth: [] }]
      summary: Apply/scrap co-occurrence recommendations (REQ-REC-002)
      responses:
        '200': { description: '{ recommendations, seed_count, scoring_factors }' }
  /api/recommendations/recompute:
    post:
      tags: [Recommendations]
      security: [{ bearerAuth: [] }]
      summary: Recompute self/school/cron (REQ-REC-005)
      responses:
        '200': { description: recomputed }
  /api/recommendations/feedback:
    post:
      tags: [Recommendations]
      security: [{ bearerAuth: [] }]
      summary: impression|click|apply (REQ-REC-006)
      responses:
        '201': { description: recorded }
  /api/field-trips:
    get:
      tags: [FieldTrips]
      summary: List field trips (school scoped, REQ-TRP-001)
      responses:
        '200': { description: '{ field_trips }' }
    post:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: Create field trip
      responses:
        '201': { description: created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT